move codes and update about-service-management.md
