+++ docs/about-service-management.md
@@ -159,7 +159,7 @@ struct EndpointParams
     int ssl_connect_timeout;
 };
 
-// In UpstreamMananger.h
+// In ServiceGovernance.h
 struct AddressParams
 {
     struct EndpointParams endpoint_params; ///< Connection config
